+++ src/Enhance/Provider.php
@@ -79,6 +79,10 @@ class Provider extends Category implements ProviderInterface
     public function create(CreateParams $params): AccountInfo
     {
         try {
+            if (!empty($params->location)) {
+                $location = $this->findLocation(trim($params->location));
+            }
+
             $plan = $this->findPlan($params->package_name);
 
             if ($customerId = $params->customer_id) {
@@ -99,7 +103,7 @@ class Provider extends Category implements ProviderInterface
             $subscriptionId = $this->createSubscription($customerId, $plan->getId());
 
             if ($domain) {
-                $this->createWebsite($customerId, $subscriptionId, $domain);
+                $this->createWebsite($customerId, $subscriptionId, $domain, $location->id ?? '');
             }
 
             return $this->getSubscriptionInfo($customerId, $subscriptionId, $domain, $email)
@@ -370,6 +374,9 @@ class Provider extends Category implements ProviderInterface
         throw $this->errorResult('Customer not found', ['email' => $email]);
     }
 
+    /**
+     * @throws ApiException
+     */
     protected function getSubscriptionInfo(
         string $customerId,
         ?int $subscriptionId,
@@ -394,6 +401,9 @@ class Provider extends Category implements ProviderInterface
             return $ns->getDomain();
         }, $this->api()->branding()->getBranding($this->configuration->org_id)->getNameServers());
 
+        $serverGroupId = $this->findServerGroupIdByWebsite($website);
+        $groupInfo = $this->findLocation($serverGroupId);
+
         return AccountInfo::create()
             ->setMessage('Subscription info obtained')
             ->setCustomerId($customerId)
@@ -405,10 +415,34 @@ class Provider extends Category implements ProviderInterface
             ->setSuspended(boolval($subscription->getSuspendedBy()))
             ->setIp($website ? implode(', ', $this->getWebsiteIps($website)) : null)
             ->setNameservers($nameservers)
+            ->setLocationId($groupInfo->id)
+            ->setLocationName($groupInfo->name)
             ->setDebug([
                 'website' => $website ? $website->jsonSerialize() : null,
                 'subscription' => $subscription->jsonSerialize(),
             ]);
+    }
+
+    /**
+     * @param Website $website
+     * @return string
+     * @throws ApiException
+     */
+    protected function findServerGroupIdByWebsite(Website $website): string
+    {
+        $servers = $this->api()->servers()->getServers();
+        $websiteServerId = $website->getAppServerId();
+
+        if (empty($servers)) {
+            throw $this->errorResult('There was an error retrieving the server information.');
+        }
+
+        foreach ($servers->getItems() as $server) {
+            if (strcasecmp($websiteServerId, $server->getId()) === 0) {
+                return $server->getGroupId();
+            }
+        }
+        throw $this->errorResult('There was an error trying to find the website server.');
     }
 
     protected function getSubscriptionUsage(
@@ -704,11 +738,18 @@ class Provider extends Category implements ProviderInterface
     /**
      * Create a new website and return the id.
      */
-    protected function createWebsite(string $customerId, int $subscriptionId, string $domain): string
+    protected function createWebsite(string $customerId, int $subscriptionId, string $domain, string $location): string
     {
-        $newWebsite = (new NewWebsite())
-            ->setSubscriptionId($subscriptionId)
-            ->setDomain($domain);
+        if (!empty($location)) {
+            $newWebsite = (new NewWebsite())
+                ->setSubscriptionId($subscriptionId)
+                ->setServerGroupId($location)
+                ->setDomain($domain);
+        } else {
+            $newWebsite = (new NewWebsite())
+                ->setSubscriptionId($subscriptionId)
+                ->setDomain($domain);
+        }
 
         return $this->api()->websites()
             ->createWebsite($customerId, $newWebsite)
@@ -930,5 +971,42 @@ class Provider extends Category implements ProviderInterface
 
         // let the provision system handle this one
         throw $e;
+    }
+
+    /**
+     * @param string $location
+     * @return object
+     * @throws ApiException
+     */
+    protected function findLocation(string $location): object
+    {
+        // Get the available groups and check the given one exist
+        $validGroupId = '';
+        $groups = $this->api()->servers()->getServerGroups();
+
+        if (empty($groups->getItems())) {
+            throw $this->errorResult('There was a problem trying to retrieve the group list');
+        }
+
+        foreach ($groups->getItems() as $group) {
+            if ($group->getId() === $location || $group->getName() === $location) {
+                $validGroup = (object)[
+                    'id' => $group->getId(),
+                    'name' => $group->getName()
+                ];
+            }
+        }
+
+        // If the input is correct, check the value against the actual server group id
+        if (!empty($validGroup)) {
+            $servers = $this->api()->servers()->getServers();
+            foreach ($servers->getItems() as $server) {
+                if ($server->getGroupId() === $validGroup->id) {
+                    return $validGroup;
+                }
+            }
+            throw $this->errorResult('There location specified is not available');
+        }
+        throw $this->errorResult('There location specified is not valid');
     }
 }